+++ packages/core/field-range/docs/0-intro.js
@@ -1,15 +1,15 @@
 // @flow
 import React from 'react';
-import { md, Example, Props } from '@atlaskit/docs';
+import { md, Example, Props, code } from '@atlaskit/docs';
 
 export default md`
   ### Usage
 
   Component which renders a slider and is a substitute of the native input[range] element
 
-  ~~~js
-  import FieldTextArea, { FieldTextAreaStateless } from '@atlaskit/field-text-area';
-  ~~~
+${code`
+import FieldTextArea, { FieldTextAreaStateless } from '@atlaskit/field-text-area';
+`}
 
   The onChange prop provides a way to subscribe to changes in the value.