+++ packages/editor/editor-core/examples/5-full-page.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 import * as React from 'react';
 import Button, { ButtonGroup } from '@atlaskit/button';
-import { akColorN300 } from '@atlaskit/util-shared-styles';
+import { akColorN90 } from '@atlaskit/util-shared-styles';
 
 import Editor from './../src/editor';
 import EditorContext from './../src/ui/EditorContext';
@@ -36,7 +36,7 @@ export const TitleInput: any = styled.input`
   padding: 0;
 
   &::placeholder {
-    color: ${akColorN300};
+    color: ${akColorN90};
   }
 `;
 TitleInput.displayName = 'TitleInput';