+++ packages/editor/editor-core/src/plugins/table/ui/TableFloatingControls/styles.ts
@@ -2,12 +2,15 @@ import styled from 'styled-components';
 // @ts-ignore: unused variable
 // prettier-ignore
 import { HTMLAttributes, ClassAttributes, ButtonHTMLAttributes, ComponentClass } from 'react';
+
 import {
-  akEditorTableToolbarSelected,
-  akEditorTableBorder,
-  akEditorTableBorderSelected,
-  akEditorTableToolbar,
-} from '../../../../styles';
+  tableToolbarSelectedColor,
+  tableToolbarColor,
+  tableBorderColor,
+  tableBorderSelectedColor,
+  tableBorderDeleteColor,
+  tableToolbarDeleteColor,
+} from '../styles';
 
 export const toolbarSize = 11;
 
@@ -15,7 +18,6 @@ export const Container: ComponentClass<HTMLAttributes<{}>> = styled.div`
   position: relative;
 
   .ProseMirror.table-resizing & {
-    top: 28px;
     display: none;
   }
   .ProseMirror.table-resizing .with-controls & {
@@ -26,17 +28,21 @@ export const Container: ComponentClass<HTMLAttributes<{}>> = styled.div`
 export const HeaderButtonDefault: ComponentClass<
   ButtonHTMLAttributes<{}>
 > = styled.button`
-  background: ${akEditorTableToolbar};
-  border-top: 1px solid ${akEditorTableBorder};
-  border-left: 1px solid ${akEditorTableBorder};
+  background: ${tableToolbarColor};
+  border-top: 1px solid ${tableBorderColor};
+  border-left: 1px solid ${tableBorderColor};
   display: block;
   padding: 0;
   cursor: pointer;
   &:hover,
   .active > &,
   .tableHovered & {
-    background-color: ${akEditorTableToolbarSelected};
-    border-color: ${akEditorTableBorderSelected};
+    background-color: ${tableToolbarSelectedColor};
+    border-color: ${tableBorderSelectedColor};
+  }
+  .danger > & {
+    background-color: ${tableToolbarDeleteColor};
+    border-color: ${tableBorderDeleteColor};
   }
   &:focus {
     outline: none;
@@ -64,18 +70,18 @@ export const InsertButtonDefault: ComponentClass<
 export const InsertMarkerDefault: ComponentClass<
   HTMLAttributes<{}>
 > = styled.div`
-  background-color: ${akEditorTableBorder};
+  background-color: ${tableBorderColor};
   position: absolute;
   height: 4px;
   width: 4px;
   border-radius: 50%;
   div:hover > & {
-    background-color: ${akEditorTableBorderSelected};
+    background-color: ${tableBorderSelectedColor};
   }
 `;
 
 export const LineMarkerDefault: ComponentClass<HTMLAttributes<{}>> = styled.div`
-  background: ${akEditorTableBorderSelected};
+  background: ${tableBorderSelectedColor};
   display: none;
   position: absolute;
   z-index: 1;