+++ packages/core/single-select/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @atlaskit/single-select
 
+## 4.0.3
+- [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
+  - @atlaskit/icon@11.3.0
+  - @atlaskit/field-base@9.0.3
+  - @atlaskit/droplist@5.0.3
+  - @atlaskit/button@7.2.5
+  - @atlaskit/theme@3.2.2
+  - @atlaskit/spinner@5.0.2
+  - @atlaskit/avatar@10.0.6
+  - @atlaskit/docs@3.0.4
+
+## 4.0.2
+- [patch] Export types for Single select [dd51dad](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dd51dad)
+- [none] Updated dependencies [dd51dad](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dd51dad)
+
 ## 4.0.1
 
 ## 4.0.0