+++ packages/media/media-filmstrip/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/media-filmstrip
 
+## 8.0.6
+- [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
+  - @atlaskit/icon@11.3.0
+  - @atlaskit/media-card@27.1.4
+  - @atlaskit/field-radio-group@2.0.3
+  - @atlaskit/media-test-helpers@12.0.4
+  - @atlaskit/media-core@18.0.3
+  - @atlaskit/checkbox@2.0.2
+  - @atlaskit/field-range@3.0.2
+  - @atlaskit/docs@3.0.4
+
+## 8.0.5
+- [patch] Fix issue with Filmstrip cutting Cards [c5b18db](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c5b18db)
+
 ## 8.0.4
 
 ## 8.0.3