+++ packages/search/quick-search/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/quick-search
 
+## 1.5.0
+- [minor] Add Avatar prop to quick search result types [aefb12d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/aefb12d)
+
+## 1.4.2
+- [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
+  - @atlaskit/icon@11.3.0
+  - @atlaskit/navigation@31.0.4
+  - @atlaskit/item@6.0.3
+  - @atlaskit/field-base@9.0.3
+  - @atlaskit/theme@3.2.2
+  - @atlaskit/avatar@10.0.6
+  - @atlaskit/docs@3.0.4
+  - @atlaskit/analytics@3.0.1
+
 ## 1.4.1
 
 ## 1.4.0